requirements.txt
# Core ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Dashboard
streamlit>=1.28.0

# Explainability
shap>=0.43.0

# Monitoring & Metrics
prometheus-client>=0.19.0

# MLOps
mlflow>=2.9.0

# Streaming
kafka-python>=2.0.2

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9

# Caching & Rate Limiting
redis>=5.0.0

# Drift Detection
scipy>=1.11.0
evidently>=0.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
